[Code] Fix ligatures involving the right boundary character.
The right boundary character may not exist in the font, so it's not
always possible to find it in the characters hash table. In such a case,
a ligature with that character will be indexed by character code instead
of by character.

Note: this raises the general question of non-existent characters, and
where/when exactly it makes sense to use them. I've added a FIXME at the
top of file.lisp for that.

2019-09-13  Didier Verna  <didier@didierverna.net>

	* core/src/file.lisp (%make-ligature/kerning-program): Handle
	ligatures with the right boundary character, which may not exist.