[Code] Export list maintenance.
This patch introduces redundancy in the export list maintenance, but
this is intentional. Every use of an exported symbol (as an accessor or a
restart name for instance) is duplicated, so that if I ever have to remove
a function or a condition, I can just delete a few lines in the export list
without worrying about other occurrences of the same symbol elsewhere.

2019-10-04  Didier Verna  <didier@didierverna.net>

	* core/src/file.lisp (extended-tfm): Reverse slots order.
	* core/meta.lisp (:net.didierverna.tfm): Export list maintenance.